requirements.txt
praw>=7.7,<8
anthropic>=0.40,<1
rich>=13.0,<14
python-dotenv>=1.0,<2
